allow a siapath to be joined with an empty siapath
The fuse package sometimes needs to combine root with root, allowing a
non-change to be joined to a dir seemed acceptable to me. Open for
feedback, if this change is rejected we have to rework fuse a bit but
it's not a non-starter - fuse can survive it just has more edge cases to
check.