dumpster fire
Actually, this is a pretty neat overhaul of the worker code. It
compiles, but I haven't run it at all and I'm sure it doesn't work. But
it's 3am and I want to pass this code forward.

Worker loop has been split out into its own file and completely
overhauled to be able to simultaneously support serial tasks and also
parallel tasks.

Also, a lot of the lock contention around caches and such has been
eliminated with the use of atomics. More on that to come soon.