tuning and clarification in the download code
There was a piece of download code that was confusing, because it was
using the term 'worker' when it was iterating over a list of pieces. I
updated the variable names to reflect this.

There was also a place where we were penalizing unresolved workers just
for being unresolved. I increased the penalties for workers that are
late, and elminated the generic penalty for being unresolved, which
should result in better estimates.

Finally, we weren't considering all of the same penalties for the
unresovled workers that we were for the resolved ones, and I couldn't
find any reaons not to have the same penalties existing for both checks,
so I extended the unresolved code to mirror the resovled code.